requirements.txt
elasticsearch>=8.15.1
gensim>=4.3.3
lxml>=5.3.0
numpy>=1.26.4
pymongo>=4.10.1
python-dotenv>=1.0.1
snowballstemmer>=2.2.0